+++ modules-services/service-templates/bin/keycloak-get-token.sh
@@ -15,6 +15,7 @@ source $PROV_CREDENTIALS
 
 CLIENT_ID="KeycloakOidcClient"
 CLIENT_ID="assistant"
+CLIENT_ID="ptm"
 
 
 if [[ $TPL_KEYCLOAK_PASSWORD = '' ]]; then